requirements.txt
dota2==1.1.0
python-dotenv==1.0.1
steam==1.4.4
Flask==3.0.2